requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-RESTX==1.3.0
Flask-JWT-Extended==4.5.3
marshmallow==3.20.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
requests==2.31.0
redis==5.0.1
bcrypt==4.1.2
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0

